+++ CHANGELOG.md
@@ -10,7 +10,11 @@ and commits should be formatted using [Conventional Commits](https://www.convent
 
 ### Changed
 
-- GitHub: Update various GH actions to actual versions
+- GitHub: Update various GH actions to actual versions by @stronk7 ([5851e6a](https://github.com/moodlehq/wiki-rag/commit/5851e6ab0a507d682bb4d9124dd5dd0a08b6d3cd))
+
+### Fixed
+
+- Search: Use only HyDE passages for dense search, not the original query
 
 ## [0.14.0] - 2026-04-06